\name{blueshark}
\docType{data}
\alias{blueshark}
\alias{shark}
\title{Blue Shark GPE3 Track}
\description{
  This data set comprises the results from the GPE3 software applied to a blue shark track collected from a miniPAT tag deployed off Halifax, Nova Scotia, Canada, by the Ocean Tracking Network in 2014. 
}
\usage{blueshark}
\format{A data.frame with three columns: the first contains the date-time stamps in POSIX, and the second and third have the longitude and latitude coordinates, respectively.}
\source{Ocean Tracking Network}

\keyword{datasets}