% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goswimming.R
\name{bootstrapCI}
\alias{bootstrapCI}
\title{Use a parametric bootstrap to calculate the standard errors of a model fit}
\usage{
bootstrapCI(mod, startseed = 42, nsamples = 50, nsteps = 20,
  savepath = NULL, setREs = FALSE, setjidx = TRUE, setlc = TRUE,
  usefirstlocs = FALSE, issm_args = list(), jumpsd = c(0.05, 0.05))
}
\arguments{
\item{mod}{The model (issm object) to bootstrap errors for}

\item{startseed}{The starting seed to keep track of the random generation}

\item{nsamples}{The number of samples to take for the bootstrap}

\item{savepath}{The path to save the results in (full path, including .rda ext)}
}
\description{
Use a parametric bootstrap to calculate the standard errors of a model fit
}
