% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf.R
\name{cdf}
\alias{cdf}
\title{Empirical Distribution Function of a Stressed Model}
\usage{
cdf(object, xCol = 1, wCol = 1)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric, the column of the underlying data 
of the \code{object} (\code{default = 1}).}

\item{wCol}{Numeric, the column of the scenario weights 
of the \code{object} (\code{default = 1}).}
}
\value{
The empirical distribution function (a function) of 
    the \code{xCol} component of the stressed model with weights 
    \code{wCol}. The empirical distribution function can be 
    evaluated at a vector.
}
\description{
Provides the empirical distribution function of a stressed model.
}
\seealso{
See \code{\link{plot_cdf}} for plotting the empirical 
    distribution function of the stressed model and 
    \code{\link{quantile_stressed}} for sample quantiles of 
    a stressed model.
}
\author{
Silvana M. Pesenti
}
