% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_multivariate.R
\name{cdf_stressed}
\alias{cdf_stressed}
\title{Empirical Distribution Function of a Stressed Model}
\usage{
cdf_stressed(object, xCol = "all", wCol = 1, grid, base = FALSE)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Vector or characters, (names of) the columns of the underlying data 
of the \code{object} (\code{default = 1}).}

\item{wCol}{Vector, the columns of the scenario weights 
of the \code{object} (\code{default = 1}).}

\item{grid}{Vector, or matrix of values for the underlying data
in \code{xCol} of the \code{object}.}

\item{base}{Logical, if TRUE, statistics under the baseline are also returned (default = "FALSE").}
}
\value{
The empirical distribution function (a function) of 
    the \code{xCol} component of the stressed model with weights 
    \code{wCol}. The empirical distribution function can be 
    evaluated at a vector.
}
\description{
Provides the empirical distribution function of a stressed 
    model component (random variable) under the scenario weights.
}
\examples{
     
## example with a stress on VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = c(0.9, 0.95), q_ratio = 1.05)
grid <- cbind(seq(min(x$normal), max(x$normal), length.out = 5), 
              seq(min(x$gamma), max(x$gamma), length.out = 5))
## stressed empirical distribution function
cdf_stressed(res1, xCol = 1, wCol = 1, grid = grid)

}
\seealso{
See \code{\link{plot_cdf}} for plotting the empirical 
    distribution function of the stressed model and 
    \code{\link{quantile_stressed}} for sample quantiles of 
    a stressed model.
}
\author{
Silvana M. Pesenti
}
