% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_numbers.R
\name{cell_numbers}
\alias{cell_numbers}
\title{Raster Cell Numbers}
\usage{
cell_numbers(grid, geom)
}
\arguments{
\item{grid}{RasterLayer object}

\item{geom}{polygon shapefile}
}
\value{
list with raster cell numbers for each polygon
}
\description{
Extracts the raster cell numbers of all cells intersecting with a spatial
object (polygon) sp
}
\note{
This function was adopted from Cesar Aybar (aybar1994@gmail.com)
}
