% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.R
\name{cor_stressed}
\alias{cor_stressed}
\title{Correlation a Stressed Model}
\usage{
cor_stressed(
  object,
  xCol = c(1, 2),
  wCol = "all",
  method = "pearson",
  base = TRUE
)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric or character vector, (names of) the columns of 
the underlying data 
of the \code{object} (\code{default = "all"}).}

\item{wCol}{Vector, the columns of the scenario weights 
of the \code{object} corresponding to different 
stresses (\code{default = "all"}).}

\item{method}{Character, one of \code{"pearson", "spearman", "kendall"}. Pearson by default.}

\item{base}{Logical, if \code{TRUE}, statistics under the baseline 
are also returned (\code{default = "FALSE"}).}
}
\value{
The correlation coefficient of the \code{xCol}
    components of the stressed model with weights \code{wCol} and method \code{wCol}.
}
\description{
Provides the correlation of two stressed model components
    (random variable) under the scenario weights.
}
\details{
\code{cor_stressed}: The correlation coefficient of two 
     chosen stressed model components, subject to the calculated scenario weights.
     The scorrelation coefficient of a stressed model component
     is denoted as \deqn{cor^W}

     The function \code{cor_stressed} provides stressed correlation coefficient of model
     components with different interpolations.
}
\section{Functions}{
\itemize{
\item \code{cor_stressed}: correlation coefficient of stressed model components
}}

\examples{
     
## example with a stress on VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = c(0.9, 0.95), q_ratio = 1.05)
## stressed correlation
cor_stressed(res1, xCol = c(1, 2), wCol = 1)
## baseline correlation
cor(x$normal, x$gamma)

}
\seealso{
See \code{\link{stress_moment}} stressing a baseline 
    model with desired moment constraints, and \code{\link{var_stressed}} provides
    stressed variances under the scenario weights
}
\author{
Kent Wu
}
