% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credit_data.R
\docType{data}
\name{credit_data}
\alias{credit_data}
\title{Credit data set}
\format{A data frame with 100,000 rows and 7 variables:
\describe{
  \item{L}{total aggregate loss of a portfolio consisting
           of three homogeneous sub-portfolios L1, L2 and L3}
  \item{L1}{aggregate loss of sub-portfolio 1}
  \item{L2}{aggregate loss of sub-portfolio 2}
  \item{L3}{aggregate loss of sub-portfolio 3}
  \item{H1}{(conditional) default probability of sub-portfolio 1}
  \item{H2}{(conditional) default probability of sub-portfolio 2}
  \item{H3}{(conditional) default probability of sub-portfolio 3}
}}
\source{
see ?
}
\usage{
credit_data
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds. conditionally binomial credit model
}
\keyword{datasets}
