% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_fast.R
\name{extract_fast}
\alias{extract_fast}
\title{Extract Raster Values Given Cell Numbers}
\usage{
extract_fast(grid, cells, fun = mean, na.rm = T)
}
\arguments{
\item{grid}{raster object}

\item{cells}{list with raster cell numbers for
each polygon given by \code{cell_numbers} function}

\item{fun}{An R function}

\item{na.rm}{logical, argument for fun}
}
\value{
a vector with mean areal values of input data for each subbasin
}
\description{
Extracts the raster values of all cells intersecting with a spatial
object (polygon) sp given by \code{cell_numbers} function and applies R function fun.
}
\note{
This function was adopted from Cesar Aybar (aybar1994@gmail.com)
}
