% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goswimming.R
\name{fit_issm}
\alias{fit_issm}
\title{fit a switching ssm iteratively}
\usage{
fit_issm(
  obs,
  ts,
  p_start_hmm,
  init_psi,
  res = c("x"),
  maxsteps = 10,
  fixsteps = FALSE,
  allowfc = FALSE,
  jiggle_fc = 0.01,
  ssm_map = NULL,
  allsilent = TRUE,
  scaleobs = 1,
  ...
)
}
\arguments{
\item{obs}{the data}

\item{ts}{the time step}

\item{p_start_hmm}{starting values for the hmm step}

\item{init_psi}{starting values for the measurement error parameter}

\item{res}{random effect values}

\item{maxsteps}{max number of steps}

\item{fixsteps}{whether to fix the number of steps, should really be true always}

\item{allowfc}{whether to allow false convergence in the ssm step}

\item{jiggle_fc}{how much random noise to add to the starting values of the parameters if the ssm falsely covnerges}

\item{ssm_map}{mapping list for the ssm step}

\item{allsilent}{if true, don't print the tmb tape for the hmm or ssm}

\item{scaleobs}{value to scale the observations by (literally just divided), helpful if they are in eastings/northings}

\item{...}{further arguments to fit_ssm}
}
\description{
fit a switching ssm iteratively
}
