% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gentrack.R
\name{genTrack}
\alias{genTrack}
\title{Simulate Correlated Random Walk Data}
\usage{
genTrack(
  nx = 500,
  ny = 1500,
  date1 = Sys.time(),
  ts = 3,
  start_seed = 42,
  process_pars = list(theta = c(0, pi), gamma = c(0.8, 0.3), sdx = rep(0.1, 2)),
  alpha = matrix(c(0.95, 0.05, 0.05, 0.95), nrow = 2, ncol = 2),
  me = "t",
  acprob = NULL,
  psi = rep(0.3, 2),
  res = NULL,
  lc = NULL,
  jidx = NULL,
  trackid = NULL,
  obsdates = NULL
)
}
\arguments{
\item{nx}{Number of simulated true locations.}

\item{ny}{Number of simulated observatioons, ny does not necessarily = nx.}

\item{date1}{Starting date.}

\item{ts}{Time step.}

\item{start_seed}{Starting seed.}

\item{process_pars}{List of relevant parameters.}

\item{alpha}{Matrix of switching probabilities. Determines the number of states you wish to have.}

\item{me}{Measurement error distribution, Gaussian or t.}

\item{acprob}{Vector of relative proportions for drawing the numbers of observed locations from each Argos class}

\item{start_loc}{Starting location in c(lon, lat). If NULL assumed to be (0,0).}
}
\value{
Simulated data set.
}
\description{
Simulate data from a DCRWS movement process with or without measurement error.
}
