% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.data}
\alias{get.data}
\title{Extracting Data}
\usage{
get.data(object, xCol = "all")
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Integer vector, columns of \code{x}, that data of the 
\code{object}, for which the summary should be returned,
(\code{default = "all"})}
}
\value{
A data.frame containing the realisations of the
        random variables on which the object of class \code{SWIM}    
        is based on.
}
\description{
Extracting data, \code{x}, from an \code{object} of class \code{SWIM}.
}
\author{
Silvana M. Pesenti
}
