% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.specs}
\alias{get.specs}
\title{Extracting Information of a Stressed Model}
\usage{
get.specs(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}
}
\value{
A data.frame containing the specifications of the \code{object}
        of class \code{SWIM}. Rows corresponds to different stresses.
        See \code{\link{SWIM}} for details.
}
\description{
Extracting the specifications of the stresses of an object 
of class \code{SWIM},
}
\seealso{
Other extracting functions: \code{\link{get.data}},
  \code{\link{get.weightsfun}}, \code{\link{get.weights}}
}
\author{
Silvana M. Pesenti
}
\concept{extracting functions}
