% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.weights}
\alias{get.weights}
\title{Extracting Weights}
\usage{
get.weights(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}
}
\value{
A data.frame containing the scenario weights, \code{new_weights} 
        of the object of class \code{SWIM}. Colums corresponds to 
        different stresses.
}
\description{
Extracting weights, \code{new_weights}, from an \code{object} of 
    class \code{SWIM}.
}
\author{
Silvana M. Pesenti
}
