% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goswimming.R
\name{get_parms}
\alias{get_parms}
\title{get the appropriate parameter values for each step, used in the issm model}
\usage{
get_parms(prev_hmm, prev_ssm, curr_model)
}
\arguments{
\item{prev_hmm}{the previous hmm}

\item{prev_ssm}{the previous ssm}

\item{curr_model}{what's next? an hmm or ssm}
}
\description{
get the appropriate parameter values for each step, used in the issm model
}
