% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipdata.R
\name{getAE}
\alias{getAE}
\title{get the Argos error terms for the measurement equation}
\usage{
getAE(ac, gpsdof = 30, gpsdiv = 10)
}
\arguments{
\item{ac}{vector of argos classes for the observations}
}
\description{
get the Argos error terms for the measurement equation
}
