% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipdata.R
\name{getJidx}
\alias{getJidx}
\title{get the weighting factors for measurement error equation
 @export}
\usage{
getJidx(dat, ts)
}
\arguments{
\item{dat}{a data frame with  date column in posix}

\item{ts}{time step in hours}
}
\description{
get the weighting factors for measurement error equation
 @export
}
