\name{greyseal}
\docType{data}
\alias{greyseal}
\alias{seal}
\title{Grey Seal GPS Track}
\description{
  This data set includes the GPS track of a grey seal tagged with a Sea Mammal Research Unit Satellite Relay Data Logger at Kouchibouguac National Park, New Brunswick, Canada, by the Department of Fisheries and Oceans in collaboration with Dalhousie University in 2013. 
}
\usage{greyseal}
\format{A data.frame with three columns: the first contains the date-time stamps in POSIX, and the second and third have the longitude and latitude coordinates, respectively.}
\source{Department of Fisheries and Oceans, Canada, in collaboration with Dalhousie University}

\keyword{datasets}