\name{laketrout}
\docType{data}
\alias{laketrout}
\alias{trout}
\alias{fish}
\title{Lake Trout VPS Track}
\description{
  This track of a lake trout was collected by a Vemco Positioning System deployed off Drummond Island in northern Lake Huron, Michigan, USA, by the Great Lakes Acoustic Telemetry Observation System in 2014. 
}
\usage{laketrout}
\format{A data.frame with three columns: the first contains the date-time stamps in POSIX, and the second and third have the longitude and latitude coordinates, respectively.}
\source{Great Lakes Acoustic Telemetry Observation System}

\keyword{datasets}