% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land.use.reclass.R
\name{land.use.reclass}
\alias{land.use.reclass}
\title{Land Use Reclassification - SWIM Hydrotope Table}
\usage{
land.use.reclass(file1, file2)
}
\arguments{
\item{file1}{character, path of land use table for reclassification}

\item{file2}{character, path of \code{'*.str'} file}
}
\description{
This function reclassifies the land use values by SWIM land use IDs in the
hydrotope table, the \code{'*.str'} file
}
\note{
The land use table should contain the \code{raster} and \code{swim} fields
}
\examples{
 \dontrun{
file1 <- "C:/grassdata/huancane/huan1/huan1_s/input/huan.str"
file2 <- "C:/PIK_2018/SWIM_PROJECT/Land_use/land_use_table_ESAC_2_SWIM.csv"
land.use.reclass(file1, file2)
}

}
