% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean_stressed}
\alias{mean_stressed}
\title{Mean of a Stressed Model}
\usage{
mean_stressed(object, xCol = 1, wCol = 1)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric or character, (name of) the column of the underlying data 
of the \code{object} (\code{default = 1}).}

\item{wCol}{Numeric, the column of the scenario weights 
of the \code{object} (\code{default = 1}).}
}
\value{
The mean of the \code{xCol}
    component of the stressed model with weights \code{wCol}.
}
\description{
Provides the mean of a stressed model component (random variable) under the scenario weights.
}
\details{
\code{mean_stressed}: The mean of a chosen stressed model component, subject to the calculated scenario weights.
     The mean of a stressed model component is denoted as \deqn{\overline{X^W}}
}
\section{Functions}{
\itemize{
\item \code{mean_stressed}: mean of stressed model components
}}

\examples{
     
## example with a stress on VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = c(0.9, 0.95), q_ratio = 1.05)
## stressed mean
mean_stressed(res1, xCol = 1, wCol = 1)
## baseline mean
mean_stressed(x$normal)

}
\seealso{
See \code{\link{stress_moment}} stressing a baseline 
    model with desired moment constraints, and \code{\link{var_stressed}} computes
    stressed variances under the scenario weights
}
\author{
Kent Wu
}
