% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{merge.SWIM}
\alias{merge.SWIM}
\title{Merging Two Stressed Models}
\usage{
\method{merge}{SWIM}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{SWIM}.}

\item{...}{Additional arguments will be ignored.}
}
\value{
An object of class \code{SWIM} containing:
  \itemize{
    \item \code{x}, the data;
    \item \code{new_weights}, a list of functions, each applied
   to a column of \code{x}, thus generating the vectors of 
   scenario weights;
    \item \code{specs}, the specification of what has been
    stressed.
    \code{specs} is a data.frame consisting of \code{type},
    \code{k}, and constraints depending on the \code{type} of stress,
    see \code{\link{SWIM}} for details.
  }
}
\description{
This function is a \code{method} for an object of class 
    \code{SWIM}.
}
\details{
Merges two objects of class \code{SWIM}, that 
    are based on the same data. \cr 
    Stresses with \code{type = c("user", "moment")} are ignored.
}
\author{
Silvana M. Pesenti
}
