% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cdf.R
\name{plot_cdf}
\alias{plot_cdf}
\title{Plot Empirical Distribution Functions}
\usage{
plot_cdf(object, xCol = 1, wCol = "all", base = FALSE, n = 500,
  x_limits, displ = TRUE)
}
\description{
Plot Empirical Distribution Functions
}
