% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_hist.R
\name{plot_hist}
\alias{plot_hist}
\title{Plotting Histograms of a Stressed Model}
\usage{
plot_hist(object, xCol = 1, wCol = "all", base = FALSE, x_limits,
  displ = TRUE)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric, the column of the underlying data 
of the \code{object} (\code{default = 1}).}

\item{wCol}{Vector, the columns of the scenario weights 
of the \code{object} (\code{default = "all"}).}

\item{base}{Logical, if \code{TRUE} the baseline is also
returned (\code{default = "FALSE"}).}

\item{x_limits}{Vector, the limits of the x-axis of the plot, the 
value for the \code{xlim} function in \code{ggplot}.}

\item{displ}{Logical, if \code{TRUE} the plot is displayed, 
otherwise the data.frame for customised plotting with 
\code{ggplot} is returned (\code{default = TRUE}).}
}
\value{
If \code{displ = TRUE}, a histogram of the stressed model 
    for different stresses.
    
    If \code{displ = FALSE}, a data.frame for customised plotting with 
    \code{ggplot}. The data.frame contains: the column, \code{xCol}, 
    of the data of the stressed model, \code{stress} (the stresses)
    and \code{value} (the values). \cr 
    Denote by \code{result} the return of the function call, then
    \code{ggplot} can be called via: 
    \deqn{ggplot(result, aes(x = result[,1], w = value, stat(density)))}
    \deqn{ + geom_freqpoly(binwidth = 0.2, aes(color = factor(stress))).}
}
\description{
Plots the histogram of a stressed model for different stresses.
}
\seealso{
See \code{\link{cdf}}, \code{\link{plot_cdf}} for 
    values, plotting, respectively, of the empirical distribution 
    function of a stressed model and \code{\link{quantile_stressed}} 
    for sample quantiles of a stressed model.
}
