% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sensitivity.R
\name{plot_sensitivity}
\alias{plot_sensitivity}
\title{plots sensitivity}
\usage{
plot_sensitivity(object, xCol = "all", wCol = "all",
  type = c("Gamma", "Wasserstein", "all"), f = NULL, displ = TRUE)
}
\description{
plots sensitivity
}
