% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipdata.R
\name{regTrack}
\alias{regTrack}
\title{regularize a track in time}
\usage{
regTrack(dat, ts)
}
\arguments{
\item{dat}{data frame, needs posix time and column names 'date', 'lon', 'lat' (even if eastings/northings)}

\item{ts}{time step in hours}
}
\description{
regularize a track in time
}
