% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampling.weights}
\alias{sampling.weights}
\title{Example data set}
\format{
A data frame with 12,022 rows and 2 variables:
\describe{
\item{OBJECTID}{Spatial grid cell identifier}
\item{weight}{Spatial grid cell sampling weight}
}
}
\usage{
sampling.weights
}
\description{
Example data set
}
\keyword{datasets}
