% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.R
\name{sd_stressed}
\alias{sd_stressed}
\title{Standard deviation and variance of a Stressed Model}
\usage{
sd_stressed(object, xCol = 1, wCol = 1)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric or character, (name of) the column of the underlying data 
of the \code{object} (\code{default = 1}).}

\item{wCol}{Numeric, the column of the scenario weights 
of the \code{object} (\code{default = 1}).}
}
\value{
The standard deviation and variance of the \code{xCol}
    component of the stressed model with weights \code{wCol}.
    Both quantities can be evaluated at a vector.
}
\description{
Provides the standard deviation and variance of a stressed 
    model component (random variable) under the scenario weights.
}
\details{
\code{sd_stressed}: The standard deviation of 
     a chosen stressed model component, subject to the calculated scenario weights.
     The standard deviation of a stressed model component
     is denoted as \deqn{sd^W}.

     The function \code{sd_stressed} provides stressed standard deviation, whereas
     the function \code{var_stressed} calculates stressed variance of model
     components with different interpolations.
}
\section{Functions}{
\itemize{
\item \code{sd_stressed}: Standard deviation and var_stressed
variance of stressed model components
}}

\examples{
     
## example with a stress on VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = c(0.9, 0.95), q_ratio = 1.05)
## stressed standard deviation
sd_stressed(res1, xCol = 1, wCol = 1)
## baseline standard deviation
sd(x$normal)

}
\seealso{
See \code{\link{stress_moment}} stressing a baseline 
    model with desired moment constraints.
}
\author{
Kent Wu
}
