% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil.reclass.R
\name{soil.reclass}
\alias{soil.reclass}
\title{Soil Reclassification - SWIM Hydrotope Table}
\usage{
soil.reclass(file1, file2)
}
\arguments{
\item{file1}{character, path of \code{swim_reclassification.dat} file}

\item{file2}{character, path of \code{'*.str'} file}
}
\description{
This function replaces the mapping units of HWSD by SWIM soil ID in the
hydrotope table, the \code{'*.str'} file
}
\examples{
 \dontrun{
file1 <- "C:/grassdata/huancane/huan1/huan1_s/input/huan.str"
file2 <- "C:/PIK_2018/SWIM_PROJECT/hwsd2swim/output/swim_reclassification.dat"
soil.reclass(file1, file2)
}
}
