% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hwsd_soil2swim.R
\name{soil2swim}
\alias{soil2swim}
\title{Harmonized World Soil Database (HWSD) to SWIM}
\usage{
soil2swim(input)
}
\arguments{
\item{input}{list of arguments: \cr  \cr
\code{hwsd.path} \cr character, location path of the 'HWSD.sqlite' file \cr \cr
\code{hwsd.soil.data} \cr character, table name that contain soil information, default value is 'HWSD_DATA'  \cr \cr
\code{hru.path} \cr character, '*.str' file path (hydrotope table) \cr \cr
\code{hru.soil.col} \cr integer, column number with soil mapping units in '*.str' file (default is 3) \cr \cr
\code{mu.number} \cr vector, soil mapping units if 'hru.path' and 'hru.soil.col' are not specified \cr \cr
\code{ptf} \cr        integer, define the method of pedotransfer function (PTF) to compute the soil hydraulic properties \cr
             1 (2) use PTFs for European (Tropical) soils \cr \cr
\code{layer_100} \cr vector, soil horizons in mm for mapping units with depth soils of 100 cm. \cr
                  e.g. c(10, 155, 300, 440, 580, 720, 860, 1000) \cr \cr
\code{str_100} \cr vector, define each soil horizons in topsoil (1: 0-30 cm) or subsoil (2: 30-100 cm) \cr
                e.g. c( 1, 1,  1,  2,  2,  2,  2,  2) \cr \cr
\code{layer_30} \cr vector, soil horizons in mm for mapping units with depth soils of 30 cm. \cr
                 e.g. c(10, 150, 300) \cr \cr
\code{str_30} \cr vector, define each soil horizons in topsoil only (1) \cr
                e.g. c( 1,  1,  1) \cr \cr
\code{layer_10} \cr vector, soil horizons in mm for mapping units with depth soils of 10 cm. \cr
                 e.g. c(10, 100) \cr \cr
\code{str_10}  \cr vector, define each soil horizons in topsoil only (1) \cr
                e.g. c( 1,  1) \cr \cr
\code{arable} \cr  integer, define the arable horizons \cr \cr
\code{output.path} \cr  character, path of output directory}
}
\value{
soil2swim writes 'non-soil_mappingunits.dat', 'swim_reclassification.dat'
and soil files for SWIM model (e.g. 'soil01.dat', 'soil02.dat', ...) in the 'output.path'.
Besides, it returns a data frame with the soil hydraulic properties.
}
\description{
This function computes the soil hydraulic properties from easily available soil properties
using pedotransfer functions (PTFs) for the SWIM model. Then, it writes the soil properties
to SWIM soil files. For these purposes, the basin soil mapping units (MU)
from HWSD raster map and the database of soil infomation of HWSD are needed. \cr

HWSD (FAO/IIASA/ISRIC/ISSCAS/JRC, 2012) has a global coverage, made up of five soil databases.
Refer to Nachtergaele et al. (2009) for further information.  \cr
Based on sand, silt, clay, reference bulk density and organic carbon soil parameters
from HWSD, the other soil properties such as porosity, available water capacity,
field capacity, saturated conductivity, soil erodibility factor K(usle) and
organic nitrogen (N) content are computed using PTFs for two soil horizons,
the topsoil (0-30 cm) and subsoil (30-100 cm).  \cr

Erodability factor K(usle) es computed according to
Wischmeier and Smith (1962) (reviewed in Romkens (1985)). \cr

Currently, there are two sets of PTFs implemented:
PTF for European soils according to Wosten et al. (2001) and Schaap  van Genuchten (2006)
and PTF for tropical soils according to Minasny and Hartemink (2011).
}
\examples{
library(DBI)
library(RSQLite)
library(dplyr)

input <- list(
hwsd.path       = "C:/PIK_2018/SWIM_PROJECT/hwsd2swim/HWSD.sqlite",
hwsd.soil.data  = "HWSD_DATA",
hru.path        = "C:/grassdata/huancane/huan1/huan1_s/input/huan.str",
hru.soil.col    = 3,
mu.number       = NULL,
ptf             = 2,
layer_100       = c(10, 155, 300, 440, 580, 720, 860, 1000),
str_100         = c( 1, 1, 1, 2, 2, 2, 2, 2),
layer_30        = c(10, 155, 300),
str_30          = c( 1, 1, 1),
layer_10        = c(10, 100),
str_10          = c(1, 1),
arable          = 5,
output.path = "C:/PIK_2018/SWIM_PROJECT/hwsd2swim/cafp"
)

soil2swim(input)

}
\references{
Budiman Minasny, Alfred E. Hartemink,
Predicting soil properties in the tropics,
Earth-Science Reviews,
Volume 106, Issues 1–2,
2011,
Pages 52-62,
https://doi.org/10.1016/j.earscirev.2011.01.005. \cr \cr


FAO/IIASA/ISRIC/ISSCAS/JRC, 2012. Harmonized World Soil Database (version 1.2).
FAO, Rome, Italy and IIASA, Laxenburg, Austria. \cr  \cr

M. Romkens. Soil erodibility factor: A perspective. Soil Erosion and Conservation.
Soil Conservation Society of America, Ankeny, Iowa. 1985. p 445-461,  1 fig, 5 tab, 31 ref., 1985.
\cr

M. G. Schaap and M. T. van Genuchten. A modified mualem-van genuchten formulation for
improved description of the hydraulic conductivity near saturation. Vadose Zone Journal, 5(1):
27, 2006. ISSN 1539-1663. doi: 10.2136/vzj2005.0005.
\cr


Nachtergaele, H. van Velthuizen, and L. Verelst. Harmonized world soil database (version 1.1).
Technical report, FAO, 2009.
\cr

J.H.M. Wösten, Ya.A. Pachepsky, W.J. Rawls,
Pedotransfer functions: bridging the gap between available basic soil data and missing soil hydraulic characteristics,
Journal of Hydrology,
Volume 251, Issues 3–4,
2001,
Pages 123-150,
https://doi.org/10.1016/S0022-1694(01)00464-4.  \cr


W. Wischmeier and D. Smith. Soil-loss estimation as a tool in soil and water management planning.
International Association Science Hydrology Comm. on Land Erosion, 59:146-159, 1962.
\cr
}
\author{
Carlos Fernandez, \email{cafpxl@gmail.com}, \email{palomino@pik-potsdam.de}, 2018 (first draft) \cr \cr
Initial script development: Stefan Ludtke (sluedtke@gfz-potsdam.de)
and Michel Wortmann (wortmann@pik-potsdam.de)
}
