% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipdata.R
\name{splitTrack}
\alias{splitTrack}
\title{split a track for a single individual based on temporal gaps}
\usage{
splitTrack(dat, cutoff = 24)
}
\arguments{
\item{dat}{the data, must have a date column in posix}

\item{the}{cutoff gap time in hours. Any gaps larger than this will force the track to be split.}
}
\description{
split a track for a single individual based on temporal gaps
}
