% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamgage.R
\name{streamgage}
\alias{streamgage}
\title{Spatial Points of Streamgage}
\usage{
streamgage(file, x, y, from_crs = "+init=epsg:4326", to_crs = NULL)
}
\arguments{
\item{file}{character, filepath of list of stations in '*.csv' format}

\item{x}{character, field name of longitude}

\item{y}{character, field name of latitude}

\item{from_crs}{character, initial coordinate reference system (CRS) by default is
World Geodetic System (WGS84) (EPSG:4326)}

\item{to_crs}{character, target CRS if conversion between projections is needed. Default is NULL.}
}
\description{
Creates spatial point object given the point attributes in '*.csv' format.
}
\examples{

 \dontrun{

file <- "C:/PIK_2018/SWIM_PROJECT/1_DATA/1_Discharge/Lurin_flow_stations.csv"
station <- streamgage(file, x = "Longitude", y = "Latitude",
                      from_crs = "+init=epsg:4326",  to_crs = "+init=epsg:32718")
plot(station)

# open R studio in GRASS to export station as GRASS vector point
# in grass gis cmd write:
# rstudio 'Grass Mapset path' &
library(rgrass7)
library(SWIM)
file <- "C:/PIK_2018/SWIM_PROJECT/1_DATA/1_Discharge/Lurin_flow_stations.csv"
station <- streamgage(file, x = "Longitude", y = "Latitude",
                      from_crs = "+init=epsg:4326",  to_crs = "+init=epsg:32718")

writeVECT(station, "station", v.in.ogr_flags=c("o", "overwrite"), driver="ESRI Shapefile")
writeVECT(station, "station", v.in.ogr_flags=c("o", "overwrite"))

# to export as ESRI Shapefile
shapefile(station, 'station')# writes to the current working directory

}
}
