% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\name{stress}
\alias{stress}
\title{this function combines the subfunctions:
stress_VaR, stress_VaR_ES, stress_prob}
\usage{
stress(type = c("VaR", "VaR ES", "prob", "user"), ...)
}
\arguments{
\item{type}{The type of stress. A choice of "VaR", "VaR ES", "prob", "user", "moments"}

\item{...}{Arguments to be passed to subsequent funcitons}
}
\value{
A SWIM object
}
\description{
this function combines the subfunctions:
stress_VaR, stress_VaR_ES, stress_prob
}
