% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_moment.R
\name{stress_mean}
\alias{stress_mean}
\title{Stressing means}
\usage{
stress_mean(x, k, new_means, ...)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{k}{Numeric vector, the
column of \code{x} that are stressed.}

\item{new_means}{Numeric vector, same length as \code{k}, 
containing the stressed means.}

\item{...}{Additional arguments to be passed to 
\code{\link[nleqslv]{nleqslv}}.}
}
\value{
A \code{\link{SWIM}} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th colum of \code{x} generate the vectors of the
      new weights;
      \item \code{specs}, the specification of what has been
      stressed.
      The \code{specs} is a data.frame consisting of ???
    }
}
\description{
Provides scenario weights such that the random variable
   under the new scenraio weights fulfils the mean constraints and
   has minimal Kullback-Leibler divergence to the baseline random
   variable.
}
\details{

}
\seealso{
Other stress functions: \code{\link{stress_VaR_ES}},
  \code{\link{stress_VaR}}, \code{\link{stress_mean_sd}},
  \code{\link{stress_moment}}, \code{\link{stress_prob}},
  \code{\link{stress_user}}, \code{\link{stress}}
}
\concept{stress functions}
