% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_moment.R
\name{stress_mean_sd}
\alias{stress_mean_sd}
\title{Stressing mean and standard deviation}
\usage{
stress_mean_sd(x, k, new_means, new_sd, ...)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{k}{Numeric vector, the
column of \code{x} that are stressed.}

\item{new_means}{Numeric vector, same length as \code{k}, 
containing the stressed means.}

\item{new_sd}{Numeric vector, same length as \code{k}, 
containing the stressed standard deviations.}

\item{...}{Additional arguments to be passed to 
\code{\link[nleqslv]{nleqslv}}.}
}
\value{
A \code{SWIM} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th colum of \code{x} generate the vectors of 
      scenario weights;
      \item \code{specs}, the specification of what has been
      stressed.
      The \code{specs} is a data.frame consisting of ???\cr
      , see \code{\link{SWIM}} for details.
    }
}
\description{
Provides scenario weights such that the random variable
   under the new scenraio weights fulfils the mean and standard
   deviation constraints and has minimal Kullback-Leibler divergence 
   to the baseline random variable.
}
\details{

}
\references{
\insertRef{Pesenti2019reverse}{SWIM}
}
\seealso{
Other stress functions: \code{\link{SWIM}},
  \code{\link{stress_VaR_ES}}, \code{\link{stress_VaR}},
  \code{\link{stress_mean}}, \code{\link{stress_moment}},
  \code{\link{stress_prob}}, \code{\link{stress_user}},
  \code{\link{stress}}
}
\concept{stress functions}
