% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_mean_std.R
\name{stress_mean_std_w}
\alias{stress_mean_std_w}
\title{Stressing Mean and Standard Deviation}
\usage{
stress_mean_std_w(x, new_mean, new_sd, k = 1, normalise = FALSE, h = NULL)
}
\arguments{
\item{x}{A vector, matrix or data frame
containing realisations of random variables. Columns of \code{x}
correspond to random variables; OR\cr
A \code{SWIMw} object, where \code{x} corresponds to the
underlying data of the \code{SWIMw} object.}

\item{new_mean}{Numeric, the stressed mean.\cr}

\item{new_sd}{Numeric, the stressed standard deviation.\cr}

\item{k}{Numeric, the column of \code{x} that is stressed
\code{(default = 1)}.\cr}

\item{normalise}{Logical. If true, values of the columns to be stressed are linearly\cr
normalised to the unit interval.\cr}

\item{h}{Function that defines the bandwidth used in KDEs. If null,
Silverman's rule will be used.\cr}

\item{gamma}{Function that defines the gamma of the risk measure. If null,
the Expected Shortfall (ES) will be used.\cr}
}
\value{
A \code{SWIMw} object containing:
    \itemize{
      \item \code{x}, a data.frame containing the data;
      \item \code{h}, bandwidths;
      \item \code{u}, vector containing the gridspace on [0, 1];
      \item \code{lam}, vector containing the lambda's of the optimized model;
      \item \code{str.fY}, function defining the densities of the stressed component;
      \item \code{str.FY}, function defining the distribution of the stressed component;
      \item \code{str.FY.inv}, function defining the quantiles of the stressed component;
      \item \code{gamma}, function defining the risk measure;
      \item \code{new_weights}, a list of functions, that applied to
  the \code{k}th column of \code{x}, generates the vectors of scenario
  weights. Each component corresponds to a different stress;
     \item \code{type = "mean sd"};
     \item \code{specs}, a list, each component corresponds to
   a different stress and contains \code{k}, \code{new_mean} 
   and \code{new_sd}.
    }
    See \code{\link{SWIM}} for details.
}
\description{
Provides weights on simulated scenarios from a baseline stochastic
    model, such that a stressed model component (random variable) fulfills a
    constraint on its mean and standard deviation.
    Scenario weights are
    selected by constrained minimisation of the Wasserstein distance to the
    baseline model.
}
\examples{
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000),
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress_wass(type = "mean sd", x = x, k = 1,
  alpha = 0.9, new_mean=1, new_sd=0.9)
  summary.SWIM(res1)

## calling stress_RM_w directly
## stressing "gamma"
res2 <- stress_mean_std_w(x = x, alpha = 0.9,
  new_mean=2.2, new_sd=1.5, k = 2)
summary.SWIM(res2)

}
\references{
\insertRef{Pesenti2019reverse}{SWIM}\cr

    \insertRef{Pesenti2020SSRN}{SWIM}\cr

    \insertRef{Csiszar1975}{SWIM}
}
\seealso{
Other stress functions: 
\code{\link{stress_HARA_RM_w}()},
\code{\link{stress_RM_mean_std_w}()},
\code{\link{stress_RM_w}()},
\code{\link{stress_VaR_ES}()},
\code{\link{stress_VaR}()},
\code{\link{stress_mean_sd}()},
\code{\link{stress_mean}()},
\code{\link{stress_moment}()},
\code{\link{stress_prob}()},
\code{\link{stress_user}()},
\code{\link{stress_wass}()},
\code{\link{stress}()}
}
\author{
Zhuomin Mao
}
\concept{stress functions}
