% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_prob.R
\name{stress_prob}
\alias{stress_prob}
\title{Stressing Value-at-Risk}
\usage{
stress_prob(x, prob, upper, lower = NULL, k = 1)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{prob}{Numeric vector, stressed probabilties corresponding to
the intervals defined through \code{lower} and 
\code{upper}.}

\item{upper}{Numeric vector, right endpoints of the intervals.}

\item{lower}{Numeric vector, left endpoints of the intervals 
(default = NULL).}

\item{k}{Numeric, the column of \code{x} that is stressed
(default = 1).}
}
\value{
A \code{SWIM} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to the
    \code{k}th colum of \code{x} generate the vectors of scenario
    weights;
    \item \code{specs}, the specification of what has been
    stressed.
    The \code{specs} is a data.frame consisting of \code{type}, \code{k},
    \code{lower}, \code{upper} and \code{prob}. Each row correponds to a different 
    stress, see \code{\link{SWIM}} for details.
    }
}
\description{
Provides scenario weights such that the random variable
   under the scenraio weights fulfils the constraints on the 
   probability sets and has minimal Kullback-Leibler divergence to 
   the baseline random variable.
}
\details{
If \code{upper = NULL}, the intervals are consequitive and 
    \code{prob} cummulative.
}
\references{
\insertRef{Pesenti2019reverse}{SWIM}
}
\seealso{
Other stress functions: \code{\link{SWIM}},
  \code{\link{stress_VaR_ES}}, \code{\link{stress_VaR}},
  \code{\link{stress_mean_sd}}, \code{\link{stress_mean}},
  \code{\link{stress_moment}}, \code{\link{stress_user}},
  \code{\link{stress}}
}
\author{
Silvana M. Pesenti
}
\concept{stress functions}
