% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_RM.R
\name{stress_RM_w}
\alias{stress_RM_w}
\title{Stressing Risk Measure}
\usage{
stress_RM_w(
  x,
  alpha,
  q_ratio = NULL,
  q = NULL,
  k = 1,
  normalise = FALSE,
  h = NULL,
  gamma = NULL
)
}
\arguments{
\item{x}{A vector, matrix or data frame
containing realisations of random variables. Columns of \code{x}
correspond to random variables; OR\cr
A \code{SWIMw} object, where \code{x} corresponds to the
underlying data of the \code{SWIMw} object.}

\item{alpha}{Numeric, vector, the levels of the stressed risk measure (RM).\cr}

\item{q_ratio}{Numeric, vector, the ratio of the stressed RM to
the baseline RM (must be same length as \code{alpha}).\cr}

\item{q}{Numeric, vector, the stressed RM at level
\code{alpha} (must be same length as \code{alpha}).\cr}

\item{k}{Numeric, the column of \code{x} that is stressed
\code{(default = 1)}.\cr}

\item{normalise}{Logical. If true, values of the columns to be stressed are linearly
normalised to the unit interval (\code{default = FALSE}).\cr}

\item{h}{Function that defines the bandwidth used in KDE (\code{default = }
Silverman's rule).\cr}

\item{gamma}{Function that defines the gamma of the risk measure which takes in 
\code{alpha} as an argument (\code{default Expected Shortfall}).}
}
\value{
A \code{SWIMw} object containing:
    \itemize{
      \item \code{x}, a data.frame containing the data;
      \item \code{h}, bandwidths;
      \item \code{u}, vector containing the gridspace on [0, 1];
      \item \code{lam}, vector containing the lambda's of the optimized model;
      \item \code{str_fY}, function defining the densities of the stressed component;
      \item \code{str_FY}, function defining the distribution of the stressed component;
      \item \code{str_FY_inv}, function defining the quantiles of the stressed component;
      \item \code{gamma}, function defining the risk measure;
      \item \code{new_weights}, a list of functions, that applied to
  the \code{k}th column of \code{x}, generates the vectors of scenario
  weights. Each component corresponds to a different stress;
     \item \code{type = "RM"};
     \item \code{specs}, a list, each component corresponds to
   a different stress and contains \code{k}, \code{alpha}, and
   \code{q}.
    }
    See \code{\link{SWIM}} for details.
}
\description{
Provides weights on simulated scenarios from a baseline stochastic
    model, such that a stressed model component (random variable) fulfills a
    constraint on its risk measure defined by a \code{gamma} function and 
    evaluated at a given level \code{alpha}. Scenario weights are
    selected by constrained minimisation of the Wasserstein distance to the
    baseline model.
}
\details{
This function implements stresses on distortion risk measures.
    Distortion risk measures are defined by a square-integrable function
    \code{gamma} where
    \deqn{\int_0^1 gamma(u) du = 1.}
    
    The distortion risk measure for some \code{gamma} and distribution 
    \code{G} is calculated as:
    \deqn{\rho_{gamma}(G) = \int_0^1 \breve(G)(u) gamma(u) du.}

    Expected Shortfall (ES) is an example of a distortion risk measure.
    The ES at level \code{alpha} of a random variable with distribution
    function F is defined by:
    \deqn{ES_{alpha} = 1 / (1 - alpha) * \int_{alpha}^1 VaR_u d u.}
}
\examples{
\dontrun{
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000),
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress_wass(type = "RM", x = x,
  alpha = 0.9, q_ratio = 1.05)
  summary(res1)

## calling stress_RM_w directly
## stressing "gamma"
res2 <- stress_RM_w(x = x, alpha = 0.9,
  q_ratio = 1.05, k = 2)
summary(res2)
}

}
\references{
\insertRef{Pesenti2019reverse}{SWIM}\cr

    \insertRef{Pesenti2020SSRN}{SWIM}\cr

    \insertRef{Csiszar1975}{SWIM}
}
\seealso{
Other stress functions: 
\code{\link{stress_HARA_RM_w}()},
\code{\link{stress_RM_mean_std_w}()},
\code{\link{stress_VaR_ES}()},
\code{\link{stress_VaR}()},
\code{\link{stress_mean_sd}()},
\code{\link{stress_mean_std_w}()},
\code{\link{stress_mean}()},
\code{\link{stress_moment}()},
\code{\link{stress_prob}()},
\code{\link{stress_user}()},
\code{\link{stress_wass}()},
\code{\link{stress}()}
}
\author{
Zhuomin Mao
}
\concept{stress functions}
