% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_user.R
\name{stress_user}
\alias{stress_user}
\title{User Defined Stress}
\usage{
stress_user(x, new_weights, k = 1)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{new_weights}{A vector, matrix or data frame containing weights.
Colums of \code{new_weights} are interpreted to correspond to random
variables. OR\cr 
A list of function, that applied to the \code{k}th colum of \code{x} 
generate the vecotrs of the new weights. \cr
\code{new_weights} are normalised to 1.}

\item{k}{Numeric, the column of \code{x} that is stressed
(default = 1).}
}
\value{
A \code{SWIM} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th component of \code{x} generate the vectors of
      scenario weights;
      \item \code{specs}, the specification of what has been
      stressed.
      The \code{specs} is a data.frame consisting of \code{type},
      \code{k} and \code{constr = user}. Each row correponds to a 
      differentstress, see \code{\link{SWIM}} for details.
    }
}
\description{
Returns a \code{SWIM} object with scenario weights defined by the user.
}
\references{
\insertRef{Pesenti2019reverse}{SWIM}
}
\seealso{
Other stress functions: \code{\link{SWIM}},
  \code{\link{stress_VaR_ES}}, \code{\link{stress_VaR}},
  \code{\link{stress_mean_sd}}, \code{\link{stress_mean}},
  \code{\link{stress_moment}}, \code{\link{stress_prob}},
  \code{\link{stress}}
}
\concept{stress functions}
