% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_VaR.R
\name{stress_VaR}
\alias{stress_VaR}
\title{Stressing Value-at-Risk}
\usage{
stress_VaR(x, alpha, q_ratio = NULL, q = NULL, k = 1)
}
\arguments{
\item{x}{Vector, matrix, data frame, realisations of a random variable or a SWIM object.}

\item{alpha}{Numeric, vector, level of VaR}

\item{q_ratio}{Numeric, vector, ratio of new VaR to old VaR, that is q /  VaR.}

\item{q}{Numeric, vector, new VaR at level alpha}

\item{k}{Numeric, the column of x that is stressed (default = 1)}
}
\value{
A SWIM object containing x, the new weights and specifications of what has been stressed, see the `SWIM` object for details.
}
\description{
This function solves the optimisation problem: Given a random variable
  and a constraint on its VaR, it provides the distribution of the random
  variable with is closest to the input wrt the Kullback-Leibler divergence
  and which fulfils the constraints.
}
\seealso{
\code{\link[SWIM]{stress}, \link[SWIM]{stress_VaR_ES}, \link[SWIM]{stress_moment}, \link[SWIM]{stress_prob}, \link[SWIM]{stress_user}}
}
\author{
Silvana M. Pesenti
}
