% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SWIM}
\alias{summary.SWIM}
\title{Summarizing a SWIM object}
\usage{
\method{summary}{SWIM}(x, xCol = "all", wCol = "all", base = FALSE)
}
\arguments{
\item{x}{A \code{SWIM} object.}

\item{xCol}{Integer vector, columns of \code{x} for which the summary
should be returned, (\code{default = "all"})}

\item{wCol}{Integer vector, columns of \code{new_weights} for which
the summary should be returned, (\code{default = "all"})}

\item{base}{Logical, if \code{TRUE} the summary of the baseline is
returned (\code{default = FALSE}).}
}
\value{
\code{summary.SWIM} returns a list where each components
    correponds to a different stress. Every component contains the
    summary statistics of the colums of the data of the \code{SWIM}
    object, including:
    \tabular{ll}{
      \code{mean}        \cr
      \code{sd}          \cr
      \code{skewness}    \cr
      \code{ex kurtosis} \cr
      \code{1Qrt}        \cr
      \code{Median}      \cr
      \code{3Qrt}        \cr
    }
}
\description{
This funcion is a \code{\link[utils]{methods}} for an object of class 
    \code{SWIM}.
}
\details{

}
\seealso{
\code{\link{summary}}, \code{\link{SWIM}}
}
\author{
Silvana M. Pesenti
}
