% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_example_shapefiles.R
\name{sw_example_shapefiles}
\alias{sw_example_shapefiles}
\title{Get the example shapefiles}
\usage{
sw_example_shapefiles(save.files = TRUE)
}
\arguments{
\item{save.files}{If TRUE, a folder with the downloaded files will be saved to the working directory. Default is TRUE.}
}
\value{
Returns example shapfiles (class SpatVector)
sw_example_shapefiles()
ls()
}
\description{
Imports a stored shapefile as a SpatVector (terra).
}
