% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_internal_shapefiles.R
\name{sw_internal_shapefiles}
\alias{sw_internal_shapefiles}
\title{Get the example shapefiles}
\usage{
sw_internal_shapefiles(x)
}
\arguments{
\item{x}{Only option is to return all}
}
\value{
Returns a list of SpatVectors
x<-sw_internal_shapefiles('all')
x
}
\description{
Imports a stored shapefile as a SpatVector (terra).
}
