% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sw_sample.R
\name{sw_sample}
\alias{sw_sample}
\title{Swim Sample}
\usage{
sw_sample(length, wghts, option = 1)
}
\arguments{
\item{length}{The number of events}

\item{wghts}{Data frame with two columns. The first column is the identity of the grid cell. The second column is the associated sampling weight. If choosing option=3, the column of sampling weights is ignored.}

\item{option}{Either 1, 2, or 3. Default is 1. Option 1 uses weighted random assignment to distribute events over grid cells. This option returns grid cell identities. Option 2 distributes events over all grid cells proportionally according to the sampling weights. Option 3 distributes events evenly over all available grid cells, ignoring sampling weights.}
}
\description{
Sample spatial grid cells and distribute events according to grid cell sample weights
}
