% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SWIM.R
\docType{package}
\name{SWIM}
\alias{SWIM}
\alias{SWIM-package}
\title{SWIM: A package for Sensitivity Analysis}
\description{
The \code{SWIM} (Scenario Weights for Importance Measurement) package
   provides scenario weights such that a stressed random variable
   (under the new scenraio weights) fulfils probabilistic constraints 
   and has minimal Kullback-Leibler divergence to the baseline random
   variable.
}
\details{
For realisations of a random variable \code{X} and a 
    (probabilstic) stress on \code{X}, the \code{SWIM}
    package provides the scenario weights, such that \code{X} under
    the scenario weights fulfils the stress and has minimal 
    Kullback-Leibler divergence to \code{X}.
    
    For a model \code{Y = g(X)} with input vector 
    \code{X = (X1,\dots,Xn)}, a stress can be defined on an input
    or the output \code{Y}. The resulting scenario weights apply to
    the entire model \code{Y = g(X)}, providing a stressed model 
    including its probability distribution. 
 
    The \code{SWIM} package is based on the \emph{reverse 
    sensitivity framework} developed in 
    \link{https://doi.org/10.1016/j.ejor.2018.10.003}. Specifically, 
    for a random variable \code{X} with proababiltiy measure \code{P},
    the algorithm solves 
    \deqn{argmin D(P | Q) s.t. constraints on the distribution
    of X under Q.}
}
\section{Stresses}{

    Implemented functions are:
    \itemize{
      \item \code{\link{stress}} calles one of the below functions by  
    defining \code{type}
      \item \code{\link{stress_VaR}} for stressing the VaR 
    (\code{type = "VaR"})
      \item \code{\link{stress_VaR_ES}} for stressing the VaR and 
    ES jointly (\code{type = "VaR ES"})
      \item \code{\link{stress_mean}} for stressing means 
    (\code{type = "mean"})
      \item \code{\link{stress_mean_sd}} for stressing means and
    standard deviations (\code{type = "mean sd"})
      \item \code{\link{stress_moment}} for stressing moments 
    (\code{type = "moment"})
      \item \code{\link{stress_prob}} for stressing intervals 
    (\code{type = "prob"})
      \item \code{\link{stress_user}} for user defined scenario weights
    (\code{type = "user"})
    }
}

\section{A \code{SWIM} object}{

    An object of calss \code{SWIM} contains a list of:
  \itemize{
    \item \code{x}, the data;
    \item \code{new_weights}, a list of functions, that applied
  to the \code{k}th colum of \code{x} generate the vectors of 
  scenario weights;
    \item \code{specs}, the specification of what has been
  stressed.
  The \code{specs} is a data.frame consisting of \code{type},
  \code{k}, and constraints depending on the \code{type} of stress.
    \itemize{
      \item \code{type = "VaR"}: \code{alpha}, the level of the 
    stressed VaR; \code{q} the stressed VaR at level \code{alpha}.
      \item \code{type = "VaR ES"}: \code{alpha}, the level of the 
    stressed VaR; \code{q} the stressed VaR at level \code{alpha};
    \code{s} the stressed ES at level \code{alpha}.
      \item \code{type = "prob"}: \code{lower}, the left endpoints of
    the intervals; \code{upper}, the right endpoints of the intervals;
    \code{prob}, stressed probabilties corresponding to the intervals
    defined through \code{lower} and \code{upper}.
      \item \code{type = "user"}: \code{constr = user}.  
      \item \code{type = "moment"}:
      \item \code{type = "mean"}:
      \item \code{type = "mean sd"}:
    }
    }
}

\seealso{
Other stress functions: \code{\link{stress_VaR_ES}},
  \code{\link{stress_VaR}}, \code{\link{stress_mean_sd}},
  \code{\link{stress_mean}}, \code{\link{stress_moment}},
  \code{\link{stress_prob}}, \code{\link{stress_user}},
  \code{\link{stress}}
}
\concept{stress functions}
