% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_project.R
\name{swim_project}
\alias{swim_project}
\title{Creates a New SWIM Project}
\usage{
swim_project(dirname, proj_name)
}
\arguments{
\item{dirname}{character, directory path where the project will be created}

\item{proj_name}{character, project name}
}
\description{
This function creates  a new SWIM Project with base files
}
\examples{

\dontrun{
dirname <- "C:/PIK_2018/SWIM_PROJECT/PERU/Pacific_drainage"
proj_name <- "lurin"

swim_project(dirname, proj_name)
}


}
