% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watch2swim.R
\name{watch2swim}
\alias{watch2swim}
\title{WATCH Forcing Data to SWIM}
\usage{
watch2swim(ncdf, subbasins, output.path = NULL, resample = FALSE,
  resolution = 0.05, scale = 2, date.ini)
}
\arguments{
\item{ncdf}{list, pathname of WATCH ncdf files. The following climates data are required: \cr \cr
\code{tmx} \cr  Maximum (2m) Air Temperature (K) \cr \cr

\code{tmp} \cr  Mean (2m) Air Temperature (K) \cr \cr

\code{tmn} \cr  Minimum (2m) Air Temperature (K) \cr \cr

\code{rai} \cr  rainfall flux (kg/m2/s) \cr \cr

\code{sno} \cr  snowfall flux (kg/m2/s) \cr \cr

\code{rhu} \cr  relative humidity (percentage) \cr \cr

\code{rad} \cr  surface downwelling shortwave radiation flux (W/m2) \cr \cr}

\item{subbasins}{SpatialPolygonsDataFrame object of subbasins with projection}

\item{output.path}{character, path of output directory}

\item{resample}{logical, if TRUE, the spatial resolution of input files are resampled
using the nearest neighbor method to the target resolution
then cells value inside of subbasin are averaged to estimate
the mean areal of climate data for each subbasin.
If FALSE, cells value covering a subbasin are identified and averaged
to compute the mean areal of climate data for each subbasin.}

\item{resolution}{numeric, target resolution for resample process}

\item{scale}{numeric, the subbasin box is multiplied by scale to mask the input ndcf files for resample process}

\item{date.ini}{character, initial date of input data in 'YYYY-MM-DD' format}
}
\value{
watch2swim writes 'clim1.dat', 'clim2.dat' files required by SWIM model.
Furthermore, each climate data for subbasins is wrote in '*.csv' files,
The latter is returned as xts objects in a list.
The climate data units are:
Minimum (tmn), maximum (tmx) and mean (tmp) temperature in °C,
rainfall (rai) in mm/day,
snow (sno) in mm/day,
relative humidity (rhu) in percentage,
radiation (rad) in MJ/d
and precipitation (pre) is the sum of rainfal and snow in mm/day.
}
\description{
This function computes the mean areal of climate data for each subbasin.
Then, it writes the climate data to SWIM climate files.
For these purposes, the WATCH Forcing Data is used.
}
\examples{
\dontrun{
# open R studio in GRASS
# in grass gis cmd write:
# rstudio 'Grass Mapset path' &
library(rgrass7)
library(sp)
require(raster)
require(rgdal)
require(parallel)
require(ncdf4)
require(rgrass7)
library(xts)

# List of ncdf files of WATCH forcing data
ncdf <- list(
 tmx = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/tasmax_watch+wfdei_1981_2013.nc4",
tmp = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/tas_watch+wfdei_1981_2013.nc4",
 tmn = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/tasmin_watch+wfdei_1981_2013.nc4",
 rai = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/pr_gpcc_watch+wfdei_1981_2013.nc4",
 sno = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/prsn_gpcc_watch+wfdei_1981_2013.nc4",
 rhu = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/hurs_watch+wfdei_1981_2013.nc4",
 rad = "C:/PIK_2018/SWIM_PROJECT/1_DATA/2_Wheater/watch/rsds_watch+wfdei_1981_2013.nc4"
 # spr = "C:/PIK_2018/SWIM_PROJECT/ncdf_star/ncdf_star/ncdf/ps_watch+wfdei_cut_1981_2013.nc4", # opt
 # win = "C:/PIK_2018/SWIM_PROJECT/ncdf_star/ncdf_star/ncdf/wind_watch+wfdei_cut_1981_2013.nc4" # opt
 )

#read the subbasins vector object from Grass Gis project
subbasins <- rgrass7::readVECT("subbasins", plugin=NULL) # or read a polygon shapefile of subbasins
plot(subbasins)

output.path ="C:/PIK_2018/SWIM_PROJECT/PERU/Grass_peru/Pacific_drainage/Lurin/InOut/input/climate"


watch2swim_out <- watch2swim(ncdf = ncdf,
                  subbasins = subbasins,
                  output.path = output.path,
                  resample = TRUE,
                  resolution = 0.05,
                  scale = 2,
                  date.ini = "1981-01-01"
                  )

}




}
