% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goswimming.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize some observations}
\usage{
winsorize(
  obs,
  speedtol = 10,
  inflate = 10,
  upperlimscale = 1.5,
  proj = FALSE,
  returnfull = FALSE
)
}
\arguments{
\item{obs}{dataframe of observations, 'id', date', 'lon', 'lat', 'lc'}

\item{speedtol}{the speed tolerance for the animal. If "empirical" this is set to}

\item{inflate}{the multiplicative inflation for the speedtol}

\item{upperlimscale}{the threshold value determining how far back the outliers should be scaled, = upperlimscale*IQR}

\item{proj}{are the data projected}

\item{returnfull}{return the index of the outliers as well as the winsorized data?}
}
\description{
Winsorize some observations
}
