% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wstor.dat.R
\name{wstor.dat}
\alias{wstor.dat}
\title{wstor.dat file for SWIM model}
\usage{
wstor.dat(subbasins = "subbasins", output.path = NULL, plot = TRUE)
}
\arguments{
\item{subbasins}{character, name of boundary vector object of subbasins in
Grass Gis project}

\item{output.path}{character, output directory path}

\item{plot}{logical, plot the subbasins boundaries if TRUE}
}
\value{
write the \code{'wstor.dat'} file
}
\description{
This function creates the \code{'wstor.dat'} file for SWIM
}
\examples{
\dontrun{
# open R studio in GRASS
# in grass gis cmd write:
# rstudio 'Grass Mapset path' &
library(rgrass7)
library(sp)
wstor.dat(subbasins = "subbasins", output.path = NULL)
}

}
