% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTrack.R
\name{genTrack}
\alias{genTrack}
\title{Simulate Correlated Random Walk Data}
\usage{
genTrack(
  n_x = 500,
  start_seed = 42,
  movement_process = "DCRWS",
  process_pars = list(theta = c(0, pi), gamma = c(0.8, 0.3), sdx = rep(0.1, 2)),
  alpha = matrix(c(0.95, 0.05, 0.05, 0.95), nrow = 2, ncol = 2),
  start_loc = NULL
)
}
\arguments{
\item{n_x}{Number of simulated true locations.}

\item{start_seed}{Starting seed}

\item{movement_process}{Type of process model to simulate from. "DCRWS" implies the first-difference correlated random walk, while "carHMM" is the conditional autoregressive correlated random walk.}

\item{process_pars}{List of relevant parameters depending on the movement process.}

\item{alpha}{Matrix of switching probabilities. Determines the number of states you wish to have.}

\item{start_loc}{Starting location in c(lon, lat). If NULL assumed to be (0,0).}
}
\value{
Simulated data set.
}
\description{
Simulate data from either a DCRWS movement process or a conditional autoregressive (step length and turning angle) process. Can simulate with or without measurement error.
}
