% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.data}
\alias{get.data}
\title{Extracting data from an object of class \code{SWIM}.}
\usage{
get.data(x, xCol = "all")
}
\arguments{
\item{x}{A \code{SWIM} object.}

\item{xCol}{Integer vector, columns of \code{x} for which the summary
should be returned, (\code{default = "all"})}
}
\value{
A data.frame containing the realisations of the
        random variables on which the object of class \code{SWIM}    
        is based on.
}
\description{
Extracting data, \code{x}, from an object of class \code{SWIM}.
}
\author{
Silvana M. Pesenti
}
