% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.data}
\alias{get.data}
\title{Extracting Data of a Stressed Model}
\usage{
get.data(object, xCol = "all")
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Vector, the columns of the underlying data 
of the \code{object} (\code{default = "all"}).}
}
\value{
A data.frame containing the realisations of the
        random variables (stochastic model) on which the 
        object of class \code{SWIM} is based.
}
\description{
Extracting the data, realisations of the stochastic model, from 
    an object of class \code{SWIM}.
}
\seealso{
Other extracting functions: \code{\link{get.specs}},
  \code{\link{get.weightsfun}}, \code{\link{get.weights}}
}
\author{
Silvana M. Pesenti
}
\concept{extracting functions}
