% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.data}
\alias{get.data}
\alias{get.weights}
\alias{get.weightsfun}
\alias{get.specs}
\title{Extracting from a Stressed Model}
\usage{
get.data(object, xCol = "all")

get.weights(object)

get.weightsfun(object)

get.specs(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Vector, the columns of the underlying data 
of the \code{object} (\code{default = "all"}).}
}
\value{
\code{get.data}: A data.frame containing the realisations of 
        the stochastic model on which the \code{object} is based.

\code{get.weights:} A data.frame containing the scenario 
    weights of the \code{object}. Columns corresponds 
    to different stresses.

\code{get.weightsfun}: A list containing the functions, that, 
    when applied to the \code{k}th column of the data, generate the 
    scenario weights of the \code{object}.\cr
    Stresses with \code{type = c("user", "moment")} will be ignored.

\code{get.specs}: A data.frame containing the specifications 
        of the \code{object}. Rows corresponds to different stresses.
        See \code{\link{SWIM}} for details.
}
\description{
Extracting the data (realisations of the stochastic model), the 
    scenario weights, the functions generating the scenario weights, 
    or the specifications of the stress from an object of class 
    \code{SWIM}.
}
\section{Functions}{
\itemize{
\item \code{get.data}: extracting data.

\item \code{get.weights}: extracting scenario weights.

\item \code{get.weightsfun}: extracting weight functions.

\item \code{get.specs}: extracting information of the stress.
}}

\examples{
## continuing example in stress_VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = 0.9, q_ratio = 1.05)
  
## returning the underlying data
all(get.data(res1) == x)
## the scenario weights
get.weights(res1)
get.weightsfun(res1)
get.specs(res1)
                            
}
\seealso{
\code{\link{SWIM}}
}
\author{
Silvana M. Pesenti
}
