% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.data}
\alias{get.data}
\alias{get.weights}
\alias{get.weightsfun}
\alias{get.specs}
\title{Extracting from a Stressed Model}
\usage{
get.data(object, xCol = "all")

get.weights(object)

get.weightsfun(object)

get.specs(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Numeric or character vector, (names of) the columns of 
the underlying data 
of the \code{object} (\code{default = "all"}).}
}
\value{
\code{get.data}: A data.frame containing the realisations of 
        the stochastic model on which the \code{object} is based.

\code{get.weights:} A data.frame containing the scenario 
    weights of the \code{object}. Columns corresponds 
    to different stresses.

\code{get.weightsfun}: A list containing functions, which, 
    when applied to a column of the data, generate the 
    scenario weights of the \code{object}. The corresponding stressed 
    columns can be obtained via \code{get.specs}.\cr
    Use \code{\link{get.weights}} if the \code{SWIM} object only contains 
    scenario weights and not a list of functions.

\code{get.specs}: A data.frame containing specifications 
        of the stresses with each row corresponding to a different 
        stress. Only a selection of the specifications is returned; 
        however, all input variables are stored in the \code{object}.
        See also \code{\link{SWIM}}.
}
\description{
Extracting the data (realisations of the stochastic model), the 
    scenario weights, the functions generating the scenario weights, 
    or the specifications of the stress from an object of class 
    \code{SWIM}.
}
\section{Functions}{
\itemize{
\item \code{get.data}: extracting data.

\item \code{get.weights}: extracting scenario weights.

\item \code{get.weightsfun}: extracting weight functions.

\item \code{get.specs}: extracting information of the stress.
}}

\examples{
## continuing example in stress_VaR
set.seed(0)
x <- as.data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2)))
res1 <- stress(type = "VaR", x = x, 
  alpha = 0.9, q_ratio = 1.05)
  
## returning the underlying data
all(get.data(res1) == x)
## the scenario weights
get.weights(res1)
get.weightsfun(res1)
get.specs(res1)
                            
}
\seealso{
\code{\link{SWIM}}
}
\author{
Silvana M. Pesenti
}
