% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.specs}
\alias{get.specs}
\title{Extracting specification of a stress}
\usage{
get.specs(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}
}
\value{
A data.frame containing the specifications of an \code{object}
        of class \code{SWIM}. Rows corresponds to different stresses.
        See \code{\link{SWIM}} object for details.
}
\description{
Extracting the specifications of an \code{object} of class \code{SWIM},
    on which the stresses are based.
}
\author{
Silvana M. Pesenti
}
