% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.weights}
\alias{get.weights}
\title{Extracting Scenario Weights of a Stressed Model}
\usage{
get.weights(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}
}
\value{
A data.frame containing the scenario weights of the object 
        of class \code{SWIM}. Columns corresponds to different 
        stresses.
}
\description{
Extracting the scenario weights from an object of 
    class \code{SWIM}.
}
\seealso{
Other extracting functions: \code{\link{get.data}},
  \code{\link{get.specs}}, \code{\link{get.weightsfun}}
}
\author{
Silvana M. Pesenti
}
\concept{extracting functions}
