% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.weights}
\alias{get.weights}
\title{Extracting weights from an object of class \code{SWIM}.}
\usage{
get.weights(x)
}
\arguments{
\item{x}{A \code{SWIM} object.}
}
\value{
A data.frame containing the scenario weights of the object
        of class \code{SWIM}. Colums corresponds to different stresses.
}
\description{
Extracting weights, \code{new_weights}, from an object of 
    class \code{SWIM}.
}
\author{
Silvana M. Pesenti
}
