% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.weightsfun}
\alias{get.weightsfun}
\title{Extracting Weight Functions of a Stressed Model}
\usage{
get.weightsfun(object)
}
\arguments{
\item{object}{A \code{SWIM} object.}
}
\value{
A list containing the functions, that, when applied to the 
    \code{k}th column of \code{X}, generate the scenario weights 
    of the \code{object} of class \code{SWIM}.
}
\description{
Extracting the list of functions from an object of class 
    \code{SWIM}, that, when applied to the \code{k}th column of 
    \code{x}, generates the scenario weights.
}
\note{
Stress with \code{type = c("user", "moment")} will be ignored.
}
\seealso{
Other extracting functions: \code{\link{get.data}},
  \code{\link{get.specs}}, \code{\link{get.weights}}
}
\author{
Silvana M. Pesenti
}
\concept{extracting functions}
