% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{get.weightsfun}
\alias{get.weightsfun}
\title{Extracting list of weights functions}
\usage{
get.weightsfun(x)
}
\arguments{
\item{x}{A \code{SWIM} object.}
}
\value{
A data.frame containing the scenario weights of an object
        of class \code{SWIM}. Colums corresponds to different stresses.
}
\description{
Extracting the list of functions from an object of class \code{SWIM},
    that, applied to the \code{k}th colum of \code{x}, gernerates the
    scenario weights.
}
\note{
A stress with type = c("user", "moment") will be ignored.
}
\author{
Silvana M. Pesenti
}
