% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_rank.R
\name{importance_rank}
\alias{importance_rank}
\title{Importance Rank of Random Variables}
\usage{
importance_rank(object, xCol = "all", wCol = "all", type = c("Gamma",
  "Wasserstein", "all"), f = NULL)
}
\description{
Importance Rank of Random Variables
}
