% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{merge.SWIM}
\alias{merge.SWIM}
\title{Merge two \code{SWIM} objects}
\usage{
\method{merge}{SWIM}(x, y)
}
\arguments{
\item{x, y}{Objects of class \code{SWIM}.}
}
\value{
An object of class \code{SWIM} containing:
  \itemize{
    \item \code{x}, the data;
    \item \code{new_weights}, a list of functions, that applied
   to the \code{k}th colum of \code{x} generate the vectors of 
   scenario weights;
    \item \code{specs}, the specification of what has been
    stressed.
    The \code{specs} is a data.frame consisting of \code{type},
    \code{k}, and constraints depending on the \code{type} of stress,
    see \code{\link{SWIM}} object for details.
    }
}
\description{
Merge two objects of class \code{SWIM}, that are based on the same
    data \code{x}.
}
\note{
A stress with type = c("user", "moment") will be ignored.
}
\author{
Silvana M. Pesenti
}
