% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{merge.SWIM}
\alias{merge.SWIM}
\title{Merging Two Stressed Models}
\usage{
\method{merge}{SWIM}(x, y, ...)
}
\arguments{
\item{x, y}{Objects of class \code{SWIM}.}

\item{...}{Additional arguments will be ignored.}
}
\value{
An object of class \code{SWIM} containing:
  \itemize{
    \item \code{x}, a data.frame containing the data;
    \item \code{new_weights}, a list, each component corresponds to 
   a different stress and is either a vector of scenario weights or a
   function, that applied to a column of \code{x}, generates the 
   vectors of scenario weights; 
    \item \code{type}, a list, each component corresponds to a 
   different stress and specifies the type of the stress;
    \item \code{specs}, a list, each component corresponds to 
   a different stress and contains a list with the specifications 
   of what has been stressed.
  }
See \code{\link{SWIM}} for details.
}
\description{
This function is a \code{method} for an object of class 
    \code{SWIM}.
}
\details{
Merges two objects of class \code{SWIM}, that 
    are based on the same data. \cr
}
\author{
Silvana M. Pesenti
}
