% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile.SWIM}
\alias{quantile.SWIM}
\title{Stressed quantiles}
\usage{
\method{quantile}{SWIM}(x, p, xCol = "all", wCol = 1, type = c("i/n",
  "(i-1)/(n-1)", "i/(n+1)", "quantile"))
}
\description{
Stressed quantiles
}
