% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.R
\name{quantile_stressed}
\alias{quantile_stressed}
\title{Sample Quantiles of a Stressed Model}
\usage{
quantile_stressed(object, probs = seq(0, 1, 0.25), xCol = "all",
  wCol = 1, type = c("quantile", "(i-1)/(n-1)", "i/(n+1)", "i/n"))
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{probs}{Numeric vector of probabilities with values 
in \code{[0,1]} (\code{default = (0, 0.25, 
0.5, 0.75, 1)}).}

\item{xCol}{Vector, the columns of the underlying data 
of the \code{object} (\code{default = "all"}).}

\item{wCol}{Numeric, the column of the scenario weights 
of the \code{object} (\code{default = 1}).}

\item{type}{Character, one of \code{"quantile","(i-1)/(n-1)",
"i/(n+1)","i/n"}, see details below.}
}
\value{
Returns estimates of distribution quantiles for 
    scenario weight \code{wCol} at the probabilities in \code{probs}.
}
\description{
Provides the sample quantiles of a stressed model.
}
\details{


The \code{type} corresponds to the algorithm for calcualting
    the estimate of the sample quantile. \code{"quantile"} 
    corresponds to the same interpolation as 
    \code{\link[stats]{quantile}}. 
    \code{"(i-1)/(n-1)", "i/(n+1)", "i/n"} are the inverse of the empirical distribution function, using, respectively, \code{(wt - 1)/T, wt/(T+1), wt/T}, where \code{wt} is the cumulative weight and \code{T} is the total weight (usually total sample size). See 
    \code{\link[Hmisc]{wtd.quantile}} on which the function
    \code{quantile_stressed} is based.
}
\seealso{
See \code{\link[Hmisc]{wtd.quantile}} on which the function
    \code{quantile_stressed} is based. \cr
    See \code{cdf} for the empifical distribution function of 
    a stressed model.
}
