% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensitivity.R
\name{sensitivity}
\alias{sensitivity}
\title{Sensitivities of a Stressed Model}
\usage{
sensitivity(object, xCol = "all", wCol = "all", type = c("Gamma",
  "Kolmogorov", "Wasserstein", "all"), f = NULL)
}
\arguments{
\item{object}{A \code{SWIM} object.}

\item{xCol}{Vector, the columns of the underlying data 
of the \code{object} (\code{default = "all"}).}

\item{wCol}{Vector, the columns of the scenario weights 
of the \code{object} (\code{default = "all"}).}

\item{type}{Character, one of \code{"Gamma", "Kolmogorov", 
"Wasserstein", "all"}.}

\item{f}{List of functions with the same length as \code{xCol}.
If provided, the transformed data, \code{f(x)},
is considered.}
}
\value{
A data.frame containing the sensitivity measures of the 
    stressed model. Rows correspond to different random variables.  
    The last two rows specify the  
    \code{stress} and \code{type} of the sensitivity measure.
}
\description{
Provides different sensitivity measures that compare the stressed 
    and the baseline model.
}
\details{
Provides sensitivity measures that compare the stressed and 
    the baseline model. Implemented sensitivity measures:
    \enumerate{
    \item 
      \code{Gamma}, the \emph{Reverse Sensitivity Measure}, defined 
      for a random variable \code{X} and scenario weights \code{w} by 
      \deqn{Gamma = ( E(X * w) - E(X) ) / normalised,}
      where the normalisation is such that \code{-1 <= |"Gamma"| <= 1}, see
      \insertCite{Pesenti2019reverse}{SWIM}. Loosely speaking, the 
      Reverse Sensitivity Measure is a (suitably normalised) difference 
      between the first moment of the stressed and the baseline 
      distribution of \code{X}.
    
    \item
      \code{Kolmogorov}, the Kolmogorov distance, defined for 
      distribution functions \code{F,G} by 
      \deqn{Kolmogorov = sup |F(x) - G(x)|.}
      Note that the Kolmogorov distance of one stress is the same for 
      all inputs. Should be used to compare different stresses.   
    
    \item
      \code{Wasserstein}, the Wasserstein distance of order 1, defined
      for two distribution functions \code{F,G} by 
      \deqn{Wasserstein = \int | F(x) - G(x)| dx.} 
    }
}
\references{
\insertRef{Pesenti2019reverse}{SWIM}
}
\seealso{
See \code{\link{importance_rank}} for ranking of random
    variables according to their sensitivities,  
    \code{\link{plot_sensitivity}} for plotting 
    sensitivity measures and \code{\link{summary}} for a 
    summary statistic of a stressed model.
}
