% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_moment.R
\name{stress_moment}
\alias{stress_moment}
\title{Stressing moments}
\usage{
stress_moment(x, f, k, m, ...)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{f}{A function, or list of functions.}

\item{k}{A vector or list of vectors indicating which columns of
\code{x} each function in \code{f} operates on.}

\item{m}{Vector of values, same length as \code{f}, containing
the new moments of \code{f(x)}. Must be in the range of
\code{f(x)}.}

\item{...}{Additional arguments to be passed to 
\code{\link[nleqslv]{nleqslv}}.}
}
\value{
A \code{SWIM} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th colums of \code{x} generate the vectors of the
      new weights;
      \item \code{specs}, the specification of what has been
      stressed.
      The \code{specs} is a data.frame consisting of ???\cr
      , see \code{\link{SWIM}} object for details.
    }
}
\description{
Provides scenario weights such that the random variable
   under the new scenraio weights fulfils the moment constraints and
   has minimal Kullback-Leibler divergence to the baseline random
   variable.
}
\details{
Calcualtes the solution wrt theta of the set of equations
}
\seealso{
Other stress functions: \code{\link{SWIM}},
  \code{\link{stress_VaR_ES}}, \code{\link{stress_VaR}},
  \code{\link{stress_mean_sd}}, \code{\link{stress_mean}},
  \code{\link{stress_prob}}, \code{\link{stress_user}},
  \code{\link{stress}}
}
\concept{stress functions}
