% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_moment.R
\name{stress_moment}
\alias{stress_moment}
\title{Stressing Moments}
\usage{
stress_moment(x, f, k, m, ...)
}
\arguments{
\item{x}{A vector, matrix or data frame 
containing realisations of random variables. Columns of \code{x} 
correspond to random variables; OR\cr
A \code{SWIM} object, where \code{x} corresponds to the 
underlying data of the \code{SWIM} object.}

\item{f}{A function, or list of functions, that applied to 
\code{x}, constitute the moment constraints.}

\item{k}{A vector or list of vectors, same length as \code{f},
indicating which columns of
\code{x} each function in \code{f} operates on.}

\item{m}{Numeric vector, same length as \code{f}, containing
the stressed moments of \code{f(x)}. Must be in the
range of \code{f(x)}.}

\item{...}{Additional arguments to be passed to 
\code{\link[nleqslv]{nleqslv}}.}
}
\value{
A \code{SWIM} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th columns of \code{x} generate the vectors of the
      new weights;
      \item \code{specs}, the specification of what has been
      stressed.
      \code{specs} is a data.frame consisting of ???
    }
    See \code{\link{SWIM}} for details.
}
\description{
Provides weights on simulated scenarios from a stochastic
    model, such that the stressed model fulfils the 
    moment constraints. Scenario weights are selected by 
    constrained minimisation of the relative entropy to the 
    baseline model.
}
\details{
The moment constraints are given by \code{E^Q( f(x) ) = m}, 
    where \code{E^Q} denotes the expectation under the stressed 
    model.\cr
    If \code{f} is a function of multiple variables, \code{k}
    corresponds to the input variables to \code{f}. Thus,
    the length of \code{k} must be equal to the number of input
    variables to \code{f}.

    The function solves the set of equations with respect to theta,
    using \code{\link[nleqslv]{nleqslv}}:
    \deqn{E^Q( f(x) ) = E( f(x) * exp(theta * f(x)) ) = m.}
}
\examples{
set.seed(0)
x <- data.frame(cbind(
  "normal" = rnorm(1000), 
  "gamma" = rgamma(1000, shape = 2), 
  "beta" = rbeta(1000, shape1 = 2, shape2 = 2)))
## stressing means
res1 <- stress(type = "mean", x = x, k = 1 : 3, 
  new_means = c(1, 1, 0.75))
## calling stress_mean directly
res2 <- stress_mean(x = x, k = 1 : 3, 
  new_means = c(1, 1, 0.75))

## stressing mean and sd of column 1
res3 <- stress(type = "mean sd", x = x, k = 1, new_means = 0.1, 
  new_sd = 1.1, method = "Newton", 
  control = list(maxit = 1000, ftol = 1E-15))
## calling stress_mean_sd directly 
res4 <- stress_mean_sd(x = x, k = 1, new_means = 0.1, 
  new_sd = 1.1, method = "Newton", 
  control = list(maxit = 1000, ftol = 1E-15))

## stressing covariance of column 1,2 while leaving the means unchanged
res5 <- stress_moment(x = x, 
  f = list(function(x)x, function(x)x, function(x)x[1] * x[2]), 
  k = list(1, 2, c(1, 2)), m = c(0, 2, 0.5), 
  method = "Newton", control = list(maxit = 1000, ftol = 1E-10))
  
## stressing jointly the VaR of column 1,3  
res6 <- stress_moment(x = x, 
  f = list(function(x)(x > 1.5), function(x)(x > 0.9)), 
  k = c(1, 3), m = c(0.9, 0.9))
     
}
\references{
\insertRef{Pesenti2019reverse}{SWIM}
}
\seealso{
See \code{\link{stress_mean}} for stressing means and 
    \code{\link{stress_mean_sd}} for stressing mean and standard 
    deviation jointly.

Other stress functions: \code{\link{stress_VaR_ES}},
  \code{\link{stress_VaR}}, \code{\link{stress_mean_sd}},
  \code{\link{stress_mean}}, \code{\link{stress_prob}},
  \code{\link{stress_user}}, \code{\link{stress}}
}
\concept{stress functions}
