% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_VaR.R
\name{stress_VaR}
\alias{stress_VaR}
\title{Stressing Value-at-Risk}
\usage{
stress_VaR(x, alpha, q_ratio = NULL, q = NULL, k = 1)
}
\arguments{
\item{x}{Vector, matrix, data frame, realisations of a random variable or a SWIM object.}

\item{alpha}{Numeric, vector, level of VaR.}

\item{q_ratio}{Numeric, vector, ratio of stressed VaR to base VaR, \eqn{q_ratio =  q /  VaR}.}

\item{q}{Numeric, vector, stressed VaR at level alpha.}

\item{k}{Numeric, the column of x that is stressed (default = 1).}
}
\value{
A `SWIM` object containing \code{x}; a list of functions, \code{new_weights}
  that, applied to the kth component of \code{x}, generate the vectors of new
  weights; and \code{specs} of what has been stressed.
  The \code{specs} are a data.frame consisting of 
      \tabular{rllll}{
        \tab \code{type} \tab \code{k} \tab \code{alpha} \tab \code{q}\cr
        stress 1 \tab VaR \tab \tab 
      }
 
   see the `SWIM` object for details.
}
\description{
This function solves the optimisation problem: Given a random variable
  and a constraint on its VaR, it provides the distribution of the random
  variable with is closest to the input wrt the Kullback-Leibler divergence
  and which fulfils the constraints.
}
\details{
The new weights are
}
\seealso{
\code{\link{stress}}, \code{\link{stress_VaR_ES}} for stressing the VaR and ES jointly, \code{\link{stress_moment}} for stressing moments, \code{\link{stress_prob}} for stressing intervals and \code{\link{stress_user}} for user defined weights.
}
\author{
Silvana M. Pesenti
}
\concept{stress functions}
