% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_VaR.R
\name{stress_VaR}
\alias{stress_VaR}
\title{Stressing Value-at-Risk}
\usage{
stress_VaR(x, alpha, q_ratio = NULL, q = NULL, k = 1)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{alpha}{Numeric vector, the levels of the stressed VaR.}

\item{q_ratio}{Numeric vector, the ratio of the stressed VaR to the
original VaR, \eqn{q_ratio =  q /  VaR}.}

\item{q}{Numeric vector, the stressed VaR at level \code{alpha}.}

\item{k}{Numeric, the column of \code{x} that is stressed
(default = 1).}
}
\value{
A \code{\link{SWIM}} object containing \code{x}, the data; \code{new_weights}, a list of functions, that applied to the \code{k}th component of \code{x} generate the vectors of the new weights; and \code{specs}, the specification of what has been stressed.
  The \code{specs} is a data.frame consisting of \code{type}, \code{k} \code{alpha} and \code{q}. Each row correponds to a different stress, see  \code{\link{SWIM}} object for details.
}
\description{
Provides the scenario weights such that the random variable
   under the new scenraio weights fulfils the constraint on the VaR and
   has minimal Kullback-Leibler divergence to the baseline random
   variable.
}
\details{
The new weights are
}
\note{
If alpha and q or q_ratio are vectors, they have to be of the same length.
}
\seealso{
\code{\link{stress}}, \code{\link{stress_VaR_ES}} for stressing the VaR and ES jointly, \code{\link{stress_moment}} for stressing moments, \code{\link{stress_prob}} for stressing intervals and \code{\link{stress_user}} for user defined weights.
}
\author{
Silvana M. Pesenti
}
\concept{stress functions}
