% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_VaR_ES.R
\name{stress_VaR_ES}
\alias{stress_VaR_ES}
\title{Stressing Value-at-Risk and Expected Shortfall}
\usage{
stress_VaR_ES(x, alpha, q_ratio = NULL, s_ratio = NULL, q = NULL,
  s = NULL, k = 1)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{alpha}{Numeric vector, the levels of the stressed VaR.}

\item{q_ratio}{Numeric vector, the ratio of the stressed VaR to the
original VaR, \eqn{q_ratio =  q /  VaR}.}

\item{s_ratio}{Numeric, vector, ratio of stressed ES to base ES,
\eqn{s_ratio = s / ES}.}

\item{q}{Numeric vector, the stressed VaR at level \code{alpha}.}

\item{s}{Numeric, vector - constraints: new ES at level alpha.}

\item{k}{Numeric, the column of \code{x} that is stressed
(default = 1).}
}
\value{
A \code{\link{SWIM}} object containing:
    \itemize{
      \item \code{x}, the data;
      \item \code{new_weights}, a list of functions, that applied to
      the \code{k}th colum of \code{x} generate the vectors of the
      new weights;
      \item \code{specs}, the specification of what has been
      stressed.
      The \code{specs} is a data.frame consisting of \code{type},
      \code{k}, \code{alpha}, \code{q} and \code{s}. Each row 
      correponds to a differentstress, see  \code{\link{SWIM}} object
      for details.
    }
}
\description{
Provides scenario weights such that the random variable
   under the new scenraio weights fulfils the constraint on the VaR and
   ES has minimal Kullback-Leibler divergence to the baseline random
   variable.
}
\details{
The new weights are 
    If q, s are vectors, they have to be of the same length.
    If q is a vector and s numeric, the stress s is used for all q's. Similarly for s vector and q numeric.
    If alpha and q or s are vectors, they have to be of the same length.
}
\seealso{
Other stress functions: \code{\link{stress_VaR}},
  \code{\link{stress_mean_sd}}, \code{\link{stress_mean}},
  \code{\link{stress_moment}}, \code{\link{stress_prob}},
  \code{\link{stress_user}}, \code{\link{stress}}
}
\concept{stress functions}
