% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_VaR_ES.R
\name{stress_VaR_ES}
\alias{stress_VaR_ES}
\title{Stressing Value-at-Risk and Expected Shortfall}
\usage{
stress_VaR_ES(x, alpha, q_ratio = NULL, s_ratio = NULL, q = NULL,
  s = NULL, k = 1)
}
\arguments{
\item{x}{Vector, matrix, data frame, realisations of a random variable or a SWIM object.}

\item{alpha}{Numeric, vector, level of VaR.}

\item{q_ratio}{Numeric, vector, ratio of stressed VaR to base VaR, \eqn{q_ratio =  q /  VaR}.}

\item{s_ratio}{Numeric, vector, ratio of stressed ES to base ES, \eqn{s_ratio = s / ES}.}

\item{q}{Numeric, vector, stressed VaR at level alpha.}

\item{s}{Numeric, vector - constraints: new ES at level alpha}

\item{k}{Numeric, the column of x that is stressed (default = 1).}
}
\description{
This function solves the optimisation problem: Given a random variable
  and a constraint on its VaR and ES, it provides the distribution of 
  the random variable with is closest to the input wrt the Kullback-Leibler
  divergence and fulfils the constraints.
}
\seealso{
\code{\link{stress}}, \code{\link{stress_VaR_ES}} for stressing the VaR and ES jointly, \code{\link{stress_moment}} for stressing moments, \code{\link{stress_prob}} for stressing intervals and \code{\link{stress_user}} for user defined weights.
}
