% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.SWIM}
\alias{summary.SWIM}
\title{Summarizing a SWIM object}
\usage{
\method{summary}{SWIM}(x, xCol = "all", wCol = "all", base = FALSE)
}
\arguments{
\item{x}{A SWIM object or a vector, matrix or data frame 
containing realisations of random variables. Colums of \code{x} 
are interpreted to correspond to random variables.}

\item{xCol}{Integer vector, columns of \code{x} for which the summary
should be returned, (\code{default = "all"})}

\item{wCol}{Integer vector, columns of \code{new_weights} for which
the summary should be returned, (\code{default = "all"})}

\item{base}{Logical, if \code{TRUE} the summary of baseline is
returned.}
}
\value{
A list...
}
\description{
This funcion is a \code{\link[utils]{methods}} for a class 
    \code{SWIM} object.
}
\details{

}
\seealso{
summary
}
\author{
Silvana M. Pesenti
}
