\name{swim-package}
\alias{swim-package}
\alias{swim}
\docType{package}
\title{Switching Movement Models}
\description{Fit a behavioral hidden Markov model to continuous animal movement data. Currently the model provided is the Hidden Markov Movement Model, the HMMM, which estimates the behavioral states along an animal track. The HMMM is appropriate for highly accurate data only.

This package additionally requires the R packages mvtnorm, sp, TMB. 
}

\author{K. Whoriskey}
