\name{swim-package}
\alias{swim-package}
\alias{swim}
\docType{package}
\title{Switching Movement Models}
\description{Fit behavioral hidden Markov models to continuous animal movement data. Currently the model provided is the Switching Hidden Markov Movement Model, the SHMMM, which estimates the behavioral state for each location of an animal track. The SHMMM is appropriate for highly accurate data only.}

\author{K. Whoriskey}
