% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goswimming.R
\name{winsorize}
\alias{winsorize}
\title{Winsorize some observations}
\usage{
winsorize(obs, inflate = 1)
}
\arguments{
\item{obs}{dataframe of observations, 'id', date', 'lon', 'lat', 'lc'}

\item{inflate}{an inflation factor for the winsorizing cut off value}
}
\description{
Winsorize some observations
}
